fix(text) move helpers to a new file
